check convert to numpy function fix

+++ bolt/src/nn/tensor/Tensor.h
@@ -54,8 +54,9 @@ class Tensor {
 
   const float* activationsPtr() const;
 
-  std::pair<std::vector<float>, std::vector<uint32_t> > topKValueIndexPair(
+  std::pair<std::vector<uint32_t>, std::vector<float> > topKValueIndexPair(
       uint32_t topk);
+
   const float* gradientsPtr() const;
 
  private:
